+++ hike-quest-app/src/components/Thread/Thread.jsx
@@ -1,10 +1,16 @@
+
 import PropTypes from 'prop-types';
-import { useContext, useState } from 'react';
+import { useContext, useState, useEffect } from 'react';
 import { AppContext } from "../../state/app.context"
 import { deleteThread, dislikeThread, likeThread } from '../../services/threads.service';
 import { useNavigate } from 'react-router-dom';
 import { getDatabase, ref, update } from 'firebase/database';
 import UpdateThreadModal from '../UpdateThreadModal/UpdateThreadModal';
+import { weatherAPI } from '../../common/constants.js'
+import ThumbsUp from '../icons/ThumbsUpOutline.jsx';
+import ThumbsUpFilled from '../icons/ThumbsUpFilled.jsx';
+
+
 
 /**
  * 
@@ -14,6 +20,7 @@ import UpdateThreadModal from '../UpdateThreadModal/UpdateThreadModal';
  *  title: string,
  *  content: string,
  *  createdOn: string,
+ *  location: string,
  *  likedBy?: string[]
  * } }} props 
  * @returns 
@@ -22,6 +29,7 @@ export default function Thread({ thread }) {
   const { userData } = useContext(AppContext);
   const navigate = useNavigate();
   const [showModal, setShowModal] = useState(false);
+  const [weatherData, setWeatherData] = useState({})
   const [currentThread, setCurrentThread] = useState(thread);
 
   const toggleLike = async () => {
@@ -41,14 +49,18 @@ export default function Thread({ thread }) {
     if (thread.author !== userData.handle && !userData.isAdmin) {
       return alert('Not authorised!');
     }
-    try {
-      await deleteThread(thread.id);
-      alert('Thread deleted successfully.');
-      navigate('/threads');
-    } catch (error) {
-      alert('Failed to delete the thread: ' + error.message);
+
+    const confirmDelete = window.confirm("Are you sure you want to delete this thread?");
+    if (confirmDelete) {
+      try {
+        await deleteThread(thread.id);
+        alert('Thread deleted successfully.');
+        navigate('/threads');
+      } catch (error) {
+        alert('Failed to delete the thread: ' + error.message);
+      }
     }
-  };
+  }
 
   const openModal = () => {
     setShowModal(true);
@@ -60,10 +72,20 @@ export default function Thread({ thread }) {
 
   const handleSubmit = (e) => {
     e.preventDefault();
+
+    const updatedThread = {
+      ...currentThread,
+    };
+
     const db = getDatabase();
-    update(ref(db, `threads/${currentThread.id}`), {
+    update(ref(db, `threads/${updatedThread.id}`, updatedThread), {
       title: currentThread.title,
       content: currentThread.content,
+
+
+
+      // location: currentThread.location,
+
     }).then(() => {
       closeModal();
     }).catch((error) => {
@@ -71,19 +93,75 @@ export default function Thread({ thread }) {
     });
   };
 
+  useEffect(() => {
+    fetch(
+      "https://api.openweathermap.org/data/2.5/weather?q=" +
+      thread.location +
+      "&units=metric&appid=" +
+      weatherAPI
+    )
+      .then((response) => response.json())
+      .then((data) => setWeatherData(data));
+  }, [])
+
+
   return (
-    <div>
+    <div className='threadContainer'>
+      <div className='userContainer'>
+        <div className='userInfo'>
+          <img
+            src="https://images.pexels.com/photos/2379004/pexels-photo-2379004.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2"
+            alt="profile-pic"
+            className="profilePic"
+          />
+          <div>
+            <p className='userName'>{thread.author}</p>
+            <p className='userType'> User type: { }</p>
+          </div>
+        </div>
+        <div>
+          {(thread.author === userData?.handle || userData?.isAdmin) && (
+            <button className="threadButtons" onClick={handleDeleteThread}>Delete</button>
+          )}
+          {thread.author === userData?.handle && <button className="threadButtons" onClick={openModal}>Edit</button>}
+
+        </div>
+      </div>
+      <p className='threadDate'> {new Date(thread.createdOn).toDateString()}</p>
       <h2 className='threadTitle'>{thread.title}</h2>
-      <p className='threadContent'>{thread.content}</p>
-      <p>Created on: {new Date(thread.createdOn).toDateString()}</p>
-      <p>Created by: {thread.author}</p>
-      <button onClick={toggleLike}>{thread.likedBy.includes(userData?.handle) ? 'Dislike' : 'Like'}</button>
-      {(thread.author === userData?.handle || userData?.isAdmin) && (
-        <>
-          <button onClick={handleDeleteThread}>Delete</button>
-        </>
+      <hr></hr>
+
+
+      {/* <p className='hashtag'> {thread.hashtag}</p> */}
+
+
+      <p className='actualThread'>{thread.content}</p>
+      <hr></hr>
+      {(weatherData?.cod === 200) && (
+        <div className='weatherContainer'>
+          <h3>Weather at {thread.location}</h3>
+          <div >
+            <img
+
+              src={`https://openweathermap.org/img/wn/${weatherData.weather[0].icon}.png`}
+            /></div>
+
+          <p>Temperature: {weatherData.main.temp} °C </p>
+
+          <p>Wind: {weatherData.wind.speed} m/s</p>
+          <p>Visibility: {weatherData.visibility} m</p>
+
+        </div>
+
       )}
-      {thread.author === userData?.handle && <button onClick={openModal}>Edit</button>}
+
+      {/* <p className='threadDate'> {thread.hashtag}</p> */}
+
+
+      <div className='buttonContainer'>
+        <button className={`threadButtons ${thread.likedBy.includes(userData?.handle) ? 'like1' : 'like0'}`} style={{ display: 'flex', alignItems: 'center' }} onClick={toggleLike}>{thread.likedBy.includes(userData?.handle) ? <ThumbsUpFilled /> : <ThumbsUp />} {thread.likedBy.length}</button>
+      </div>
+
       <UpdateThreadModal
         show={showModal}
         handleClose={closeModal}
@@ -91,7 +169,9 @@ export default function Thread({ thread }) {
         thread={currentThread}
         setThread={setCurrentThread}
       />
-    </div>
+    </div >
+
+
   );
 }
 
@@ -102,6 +182,7 @@ Thread.propTypes = {
     title: PropTypes.string.isRequired,
     content: PropTypes.string.isRequired,
     createdOn: PropTypes.string.isRequired,
+    location: PropTypes.string.isRequired,
     likedBy: PropTypes.arrayOf(PropTypes.string).isRequired,
   }).isRequired,
 };
